refactor: remove duplicated components, and move form components to common directory
